requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
openpyxl==3.1.2
reportlab==4.2.5
python-dateutil==2.9.0.post0
